Exhibit API: include `last_published_at` in meta
