+++ ov_wag/api.py
@@ -7,8 +7,7 @@ from wagtail.images.api.v2.views import ImagesAPIViewSet
 from wagtail_headless_preview.models import PagePreview
 
 from authors.views import AuthorsAPIViewSet
-from exhibits.api import FeaturedExhibitsAPIViewSet
-from exhibits.views import ExhibitsAPIViewSet
+from exhibits.api import ExhibitsAPIViewSet, FeaturedExhibitsAPIViewSet
 from ov_collections.views import CollectionAPIViewSet
 
 # Create the router. 'wagtailapi' is the URL namespace
